install standard library
Installs the standard platform independent libraries from Lib. Also updates
README.md to clarify how python will try to find these files, in most
cases they should be found automatically.

Also updates the README to clarify that a static executable will only
be built with the musl (gnu produces a dynamic exe) as this sentence
was a part of the docs on runtime libraries.